Update the hook documentation
It is always necessary to run easy install to get a hook to be put in place.
The old instructions were misleading, because they implied that it is
possible to add a hook for new repositories only.

Also add instructions on how to remove a hook, and why removal via gitolite
is not possible.